requirements.txt
coverage==4.5.2
Django==2.1.5
django-nose==1.4.6
djangorestframework==3.9.1
nose==1.3.7
pytz==2018.9
